# EcommerceAPI.postman_collection.json-Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "EcommerceAPI",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/register": {
      "post": {
        "summary": "register user",
        "tags": [
          "User"
        ],
        "operationId": "registeruser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/registeruserrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "summary": "login user",
        "tags": [
          "User"
        ],
        "operationId": "loginuser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/loginuserrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/getAllUser": {
      "get": {
        "summary": "get user",
        "tags": [
          "User"
        ],
        "operationId": "getuser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/4": {
      "get": {
        "summary": "get user id",
        "tags": [
          "User"
        ],
        "operationId": "getuserid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/1": {
      "put": {
        "summary": "update user",
        "tags": [
          "User"
        ],
        "operationId": "updateuser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateuserrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "summary": "detele user",
        "tags": [
          "User"
        ],
        "operationId": "deteleuser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/verifyTokenEmail/eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Mywicm9sZSI6MSwiaWF0IjoxNjI2Mjk3NTUwLCJleHAiOjE2MjYzODM5NTB9.OdEPj_ki4m4t1IYyECfNEY7jEEmCtBtKDIqyrxI4Ni8": {
      "post": {
        "summary": "verify token email",
        "tags": [
          "User"
        ],
        "operationId": "verifytokenemail",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/verifytokenemailrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/voucher/createVoucher": {
      "post": {
        "summary": "create voucher",
        "tags": [
          "voucher"
        ],
        "operationId": "createvoucher",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createvoucherrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/voucher/getVoucher": {
      "get": {
        "summary": "get voucher",
        "tags": [
          "voucher"
        ],
        "operationId": "getvoucher",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/voucher/getVoucher/1": {
      "get": {
        "summary": "get voucher id",
        "tags": [
          "voucher"
        ],
        "operationId": "getvoucherid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/voucher/updateVoucher/1": {
      "put": {
        "summary": "update voucher",
        "tags": [
          "voucher"
        ],
        "operationId": "updatevoucher",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updatevoucherrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/voucher/deleteVoucher/1": {
      "delete": {
        "summary": "delete voucher",
        "tags": [
          "voucher"
        ],
        "operationId": "deletevoucher",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
    "registeruserrequest": {
      "title": "registeruserrequest",
      "example": {
        "username": "quangdo123",
        "email": "quangdo123@gmail.com",
        "password": "12345",
        "phonenumber": 32999995,
        "address": "Hanoi"
      },
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "phonenumber": {
          "type": "integer",
          "format": "int32"
        },
        "address": {
          "type": "string"
        }
      },
      "required": [
        "username",
        "email",
        "password",
        "phonenumber",
        "address"
      ]
    },
    "loginuserrequest": {
      "title": "loginuserrequest",
      "example": {
        "email": "quangdo123@gmail.com",
        "password": "12345"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "updateuserrequest": {
      "title": "updateuserrequest",
      "example": {
        "username": "Hoa"
      },
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        }
      },
      "required": [
        "username"
      ]
    },
    "verifytokenemailrequest": {
      "title": "verifytokenemailrequest",
      "example": {
        "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Mywicm9sZSI6MSwiaWF0IjoxNjI2Mjk3NTUwLCJleHAiOjE2MjYzODM5NTB9.OdEPj_ki4m4t1IYyECfNEY7jEEmCtBtKDIqyrxI4Ni8"
      },
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "required": [
        "token"
      ]
    },
...
    "createvoucherrequest": {
      "title": "createvoucherrequest",
      "example": {
        "code": "ABC123",
        "quantity": 12,
        "discount": 10,
        "startTime": "2021-07-15T03:00:00",
        "stopTime": "2021-07-15T12:55:00"
      },
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "discount": {
          "type": "integer",
          "format": "int32"
        },
        "startTime": {
          "type": "string"
        },
        "stopTime": {
          "type": "string"
        }
      },
      "required": [
        "code",
        "quantity",
        "discount",
        "startTime",
        "stopTime"
      ]
    },
    "updatevoucherrequest": {
      "title": "updatevoucherrequest",
      "example": {
        "code": "134AGBF",
        "quantity": 20,
        "discount": 12,
        "startTime": "2021-7-5 2:22:0",
        "stopTime": "2021-7-5 4:7:0"
      },
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "discount": {
          "type": "integer",
          "format": "int32"
        },
        "startTime": {
          "type": "string"
        },
        "stopTime": {
          "type": "string"
        }
      },
      "required": [
        "code",
        "quantity",
        "discount",
        "startTime",
        "stopTime"
      ]
    }
  },
...
}